qd/requirements.txt
gym==0.18.0
numpy==1.21.4
scipy==1.7.2
pandas==1.2.3
numba==0.52.0
gin-config==0.4.0
fire==0.4.0
scikit-learn==0.24.1
cma==3.0.3
logdir==0.12.0
loguru==0.5.3
matplotlib==3.4.2
seaborn==0.12.0
alive-progress==1.6.2
cloudpickle==1.6.0
ruamel.yaml == 0.17.7
python-slugify==5.0.2
pingouin==0.4.0
Pillow==7.2.0
PyVirtualDisplay==2.2
opencv-python==4.5.5.64
requests==2.28.1

# pyribs currently installed from master version; these are its dependencies
numpy_groupies>=0.9.16
sortedcontainers>=2.0.0
threadpoolctl>=3.0.0
semantic-version>=2.10

# Mario - requires interfacing with Java.
pyjnius==1.4.1

# Overcooked - see
# https://github.com/HumanCompatibleAI/overcooked_ai/blob/master/setup.py#L26
overcooked-ai==1.1.0
ipython==7.29.0
ipywidgets==7.6.5
tqdm==4.62.3
pygame==2.0.0

# Debugging
pyinstrument==3.3.0
pytest==6.2.2
pytest-cov==2.10.1

# Dask
dask==2.30.0
distributed==2.30.0
click==7.1.2  # Newer click causes error with dask scheduler.
bokeh==2.2.3
jupyter-server-proxy==1.5.0

# Pytorch
-f https://download.pytorch.org/whl/cu113/torch_stable.html
torch==1.10.0+cu113
torchvision==0.11.1+cu113
torchaudio==0.10.0+cu113

# Maze
gym-minigrid==1.0.1
stable-baselines3==1.4.0
scikit-image==0.19.2
